added regenerate, sanitized input
